Cut communication between erizoController and nuve
We'd prefer not to run mongodb in production, and it looks as if nuve is
the only thing that talks to mongodb.  From messing around with the
basicExample, it looks as if nuve/mongo is basically a cache for info
that's either in config files or hard-coded.  This commit simplifies the
code by moving those values to the basic example client-side js.

It looks like we can do what we need to do with no nuve running,
therefore no mongdb.