Remove authorization from erizoController
It's not useful to our cut-down recording use case so over the gunwale
it goes.  The user authn wasn't really authentication anyway: the user
asserted a hard-coded name and role, and the server trusted it.  It's
easier for the time being to move the hard-coded-ness out of the
interface and into the guts of the server.  Eventually we should rip it
out completely.