Stub out room names
Since we're going to be recording individual streams we have no need for
the "room" abstraction, so for now we'll remove it from the interface.
For the time being, there's still a single room under the hood, but it's
hard-coded in the server so we can clean up the implementation later.