Install Centos' glib package instead of compiling
The version in the yum repos for Centos 7 is plenty new enough so we
don't need to download, compile, etc we can just install from yum.  In
fact, we might not need to install it explicitly since many things
depend on it, but we should probably install it explicitly in a later
commit.